switch toCamelCas
