+++ js/stateless.js/package.json
@@ -45,6 +45,7 @@
         "buffer": "6.0.3",
         "buffer-layout": "^1.2.2",
         "camelcase": "^8.0.0",
+        "camelcase-keys": "^9.1.3",
         "superstruct": "2.0.2"
     },
     "devDependencies": {
